Modify starting population of moretopo case
Hopefully this time it works. It modifies the passDomBounds to be the same
as the normal amout of topo and use origPop = 'initbusemann2' which
is a much better process for spreading fill if it goes above 1. Making sure enough
volume is added and not producing too many "Burgers".

Also Added an explicit AdaptSizeHalfBusemann update.